+++ android/src/main/kotlin/com/example/social_sharing_plus/utils/SocialConstants.kt
@@ -11,6 +11,9 @@ object SocialConstants {
     const val WHATSAPP = "shareToWhatsApp"
     const val REDDIT = "shareToReddit"
     const val TELEGRAM = "shareToTelegram"
+    const val INSTAGRAM = "shareToInstagram"
+    const val INSTAGRAM_STORIES = "shareToInstagramStories"
+    const val INSTAGRAM_REELS = "shareToInstagramReels"
 
     // PACKAGE NAMES
     const val FACEBOOK_PACKAGE_NAME = "com.facebook.katana"
@@ -19,6 +22,11 @@ object SocialConstants {
     const val WHATSAPP_PACKAGE_NAME = "com.whatsapp"
     const val REDDIT_PACKAGE_NAME = "com.reddit.frontpage"
     const val TELEGRAM_PACKAGE_NAME = "org.telegram.messenger"
+    const val INSTAGRAM_PACKAGE_NAME = "com.instagram.android"
+
+    // INSTAGRAM INTENT ACTIONS
+    const val INSTAGRAM_STORIES_INTENT = "com.instagram.share.ADD_TO_STORY"
+    const val INSTAGRAM_REELS_INTENT = "com.instagram.share.ADD_TO_REEL"
 
     // WEB URL
     const val FACEBOOK_WEB_URL = "https://www.facebook.com/sharer/sharer.php?u="
@@ -27,4 +35,5 @@ object SocialConstants {
     const val WHATSAPP_WEB_URL = "https://api.whatsapp.com/send?text="
     const val REDDIT_WEB_URL = "https://www.reddit.com/submit?title="
     const val TELEGRAM_WEB_URL = "https://t.me/share/url?url="
+    const val INSTAGRAM_WEB_URL = "https://www.instagram.com/"
 }